Update storage utils for saving and extracting persitent settings of table
