--- packages/design-system/src/components/table/useTable/usePersistentSettings/index.tsx
@@ -1,86 +1,0 @@
-/*
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * External dependencies
- */
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import extractStorage from './extractStorage';
-import updateStorage from './updateStorage';
-import { TableColumn, TableFilter } from '..';
-
-export type PersistentOptions = {
-  columns?: TableColumn[];
-  visibleColumns?: TableColumn[];
-  sortBy?: string;
-  sortOrder?: 'asc' | 'desc';
-  selectedFilters?: TableFilter;
-  searchValue?: string;
-};
-
-export type PersistentStorageData = {
-  columnsVisibility: { [key: string]: boolean };
-  columnsSizing: { [key: string]: number };
-  sortBy?: string;
-  sortOrder?: 'asc' | 'desc';
-  selectedFilters: {
-    [key: string]: TableFilter[keyof TableFilter]['filterValues'];
-  };
-  searchValue: string;
-};
-
-const usePersistentSettings = (persistenceKey?: string) => {
-  const [tableOptions, setTableOptions] = useState<PersistentStorageData>();
-
-  const isChromeExtension = useMemo(() => {
-    return window.location.protocol === 'chrome-extension:';
-  }, []);
-
-  useEffect(() => {
-    if (!persistenceKey) {
-      return;
-    }
-
-    (async () => {
-      const data =
-        (await extractStorage(persistenceKey, isChromeExtension)) || {};
-
-      setTableOptions({
-        columnsVisibility: data?.columnsVisibility,
-        columnsSizing: data?.columnsSizing,
-        sortBy: data?.sortBy,
-        sortOrder: data?.sortOrder,
-        selectedFilters: data?.selectedFilters,
-        searchValue: data?.searchValue,
-      });
-    })();
-  }, [persistenceKey, isChromeExtension]);
-
-  const updater = useCallback(
-    async (newOptions: PersistentOptions) => {
-      if (!persistenceKey) {
-        return;
-      }
-
-      await updateStorage(persistenceKey, isChromeExtension, newOptions);
-    },
-    [isChromeExtension, persistenceKey]
-  );
-
-  return [tableOptions, updater] as const;
-};
-
-export default usePersistentSettings;
